fixed another crahsbug in aai
git-svn-id: https://spring.clan-sy.com/svn/spring/trunk@2393 37977431-3df6-0310-b722-df95706aa16b

+++ AI/Global/AAI/AAIBuildTable.h
@@ -179,8 +179,8 @@ public:
 	static UnitTypeStatic *units_static;
 
 	// storage for def. building selection
-	static float **def_power;
-	static float *max_pplant_eff;
+	static double **def_power;
+	static double *max_pplant_eff;
 
 	//
 	//	non static variales
